improvements and cleanups
